Move background tasks to their own views file
