+++ core/views/__init__.py
@@ -10,3 +10,4 @@ from .field_mapper import *
 from .published import *
 from .record_identifer_transformation_scenario import *
 from .stateio import *
+from .background_tasks import *
